fix(frontend): - move POC endpoints (authinfo) under /poc/* - relocate the poc endpoints to their own directory [pulumi up]
